worker: check if path exists

+++ webscanner/worker.py
@@ -111,7 +111,8 @@ def downloader():
                 for dtest in dirtytests:
                     if dtest.percent_done() == 100:
                         log.debug("Cleanup old test %s (path: %s)"%(dtest.uuid, dtest.download_path))
-                        shutil.rmtree(dtest.download_path)
+                        if os.path.exists(dtest.download_path):
+                            shutil.rmtree(dtest.download_path)
                         dtest.is_deleted = True
                         dtest.save()
